backend-v1/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Google Cloud
google-cloud-firestore==2.13.1
google-cloud-storage==2.10.0
google-cloud-tasks==2.15.0
google-cloud-secret-manager==2.17.0
google-cloud-logging==3.8.0

# Redis (for caching and rate limiting)
redis==5.0.1

# HTTP & Async
httpx==0.25.2
aiofiles==23.2.1

# Utilities
ulid-py==1.1.0
python-dateutil==2.8.2
pytz==2023.3

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# API Documentation
python-jose==3.3.0

# RevenueCat webhook validation
cryptography==41.0.7

# Image processing (for Feng Shui analysis)
pillow==10.1.0
opencv-python-headless==4.8.1.78

# AI/ML (optional, for advanced analysis)
# google-cloud-aiplatform==1.38.1
# langchain==0.0.348
# openai==1.6.1